1.更新thymeleaf版本到3.0 2.给user表添加gender属性 3.用户个人信息展示

+++ src/main/java/com/ruoyi/project/system/post/controller/PostController.java
@@ -1,22 +1,18 @@
 package com.ruoyi.project.system.post.controller;
 
-import java.util.List;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import com.ruoyi.framework.aspectj.lang.annotation.Log;
 import com.ruoyi.framework.web.controller.BaseController;
 import com.ruoyi.framework.web.domain.Message;
 import com.ruoyi.framework.web.page.TableDataInfo;
 import com.ruoyi.project.system.post.domain.Post;
 import com.ruoyi.project.system.post.service.IPostService;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 /**
  * 岗位信息操作处理
@@ -62,6 +58,10 @@ public class PostController extends BaseController
         if (post == null)
         {
             return Message.error("岗位不存在");
+        }
+        if (postService.selectCountPostById(postId) > 0)
+        {
+            return Message.error("岗位已分配,不能删除");
         }
         if (postService.deletePostById(postId) > 0)
         {
